Added default implementation for MapBox with integrated AccessToken,StyleID, current location feature
