+++ feature/home/src/main/java/com/epmedu/animeal/home/HomeScreen.kt
@@ -1,191 +1,15 @@
 package com.epmedu.animeal.home
 
-import android.Manifest
-import android.content.pm.PackageManager
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.viewinterop.AndroidView
-import com.epmedu.animeal.base.theme.Typography
-import com.epmedu.animeal.extensions.launchAppSettings
-import com.epmedu.animeal.foundation.button.AnimealButton
-import com.epmedu.animeal.foundation.switch.AnimealSwitch
-import com.epmedu.animeal.resources.R
-import com.google.accompanist.permissions.ExperimentalPermissionsApi
-import com.google.accompanist.permissions.isGranted
-import com.google.accompanist.permissions.rememberPermissionState
-import com.google.accompanist.permissions.shouldShowRationale
-import com.google.android.gms.location.LocationServices
-import com.mapbox.geojson.Point
-import com.mapbox.maps.CameraOptions
-import com.mapbox.maps.MapInitOptions
-import com.mapbox.maps.MapView
-import com.mapbox.maps.Style
-import com.mapbox.maps.plugin.locationcomponent.location
-
-@OptIn(ExperimentalPermissionsApi::class)
-@Composable
-fun CheckLocationPermission(onGranted: @Composable () -> Unit) {
-    var isPermissionRequested by remember {
-        mutableStateOf(false)
-    }
-    val permissionState = rememberPermissionState(
-        permission = Manifest.permission.ACCESS_FINE_LOCATION,
-        onPermissionResult = {
-            isPermissionRequested = true
-        }
-    )
-
-    LaunchedEffect(Unit) {
-        permissionState.launchPermissionRequest()
-    }
-    with(permissionState.status) {
-        when {
-            isGranted -> onGranted()
-            shouldShowRationale -> RationaleText()
-            isPermissionRequested && !isGranted && !shouldShowRationale -> NavigateToSettingsPrompt()
-        }
-    }
-}
+import androidx.hilt.navigation.compose.hiltViewModel
 
 @Composable
 fun HomeScreen() {
-    CheckLocationPermission {
-        Maps()
-    }
-}
+    val homeViewModel = hiltViewModel<HomeViewModel>()
 
-@Composable
-private fun Maps() {
+    val state by homeViewModel.stateFlow.collectAsState()
 
-    Box(modifier = Modifier.fillMaxSize()) {
-        MapboxMap()
-        AnimealSwitch(
-            modifier = Modifier
-                .align(alignment = Alignment.TopCenter)
-                .padding(top = 24.dp),
-            onTabSelected = {}
-        )
-    }
-}
-
-private const val LATITUDE = 60.239
-private const val LONGITUDE = 25.004
-
-@Composable
-private fun MapboxMap() {
-    val context = LocalContext.current
-
-    val fusedLocationProvider = LocationServices.getFusedLocationProviderClient(context)
-
-
-    val mapView = mapView()
-    AndroidView(
-        factory = { mapView },
-        modifier = Modifier.fillMaxSize()
-    ) { mapView ->
-        mapView.getMapboxMap()
-            .apply {
-                loadStyleUri(
-                    Style.MAPBOX_STREETS,
-                    // After the style is loaded, initialize the Location component.
-                    object : Style.OnStyleLoaded {
-                        override fun onStyleLoaded(style: Style) {
-                            mapView.location.updateSettings {
-                                enabled = true
-                                pulsingEnabled = true
-                            }
-                        }
-                    }
-                )
-                setCamera(
-                    CameraOptions.Builder()
-                        .center(Point.fromLngLat(LONGITUDE, LATITUDE))
-                        .zoom(9.0)
-                        .build()
-                )
-            }
-    }
-
-    fusedLocationProvider.lastLocation.addOnSuccessListener {
-        mapView.getMapboxMap().setCamera(
-            CameraOptions.Builder()
-                .center(Point.fromLngLat(it.longitude, it.latitude))
-                .zoom(9.0)
-                .build()
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun DefaultPreview() {
-    MapboxMap()
-}
-
-@Composable
-private fun mapView(): MapView {
-    val context = LocalContext.current
-    val metadata = context.packageManager.getApplicationInfo(
-        context.packageName,
-        PackageManager.GET_META_DATA
-    ).metaData
-
-    val mapInitOptions = MapInitOptions(context).apply {
-        resourceOptions.toBuilder().accessToken(metadata.get("mapbox.ACCESS_TOKEN").toString())
-            .build()
-    }
-
-    return MapView(context, mapInitOptions)
-}
-
-@Composable
-private fun RationaleText() {
-    Box(modifier = Modifier.fillMaxSize()) {
-        Text(
-            text = stringResource(R.string.location_rationale_text),
-            modifier = Modifier.align(Alignment.Center)
-        )
-    }
-}
-
-@Composable
-private fun NavigateToSettingsPrompt() {
-    val context = LocalContext.current
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(24.dp),
-        verticalArrangement = Arrangement.Center
-    ) {
-        Text(
-            text = stringResource(R.string.location_permission_denied),
-            textAlign = TextAlign.Center,
-            style = MaterialTheme.typography.body1
-        )
-        AnimealButton(
-            modifier = Modifier
-                .align(Alignment.CenterHorizontally)
-                .padding(top = 24.dp),
-            text = stringResource(R.string.navigate_to_app_settings),
-            onClick = { context.launchAppSettings() }
-        )
-    }
+    HomeScreenUI(state = state)
 }
