+++ feature/home/src/main/java/com/epmedu/animeal/home/HomeViewModel.kt
@@ -1,0 +1,31 @@
+package com.epmedu.animeal.home
+
+import androidx.lifecycle.viewModelScope
+import com.epmedu.animeal.common.component.BuildConfigProvider
+import com.epmedu.animeal.common.component.LocationProvider
+import com.epmedu.animeal.common.domain.StateViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+import kotlinx.coroutines.launch
+
+@HiltViewModel
+internal class HomeViewModel @Inject constructor(
+    private val buildConfigProvider: BuildConfigProvider,
+    private val locationProvider: LocationProvider
+) : StateViewModel<HomeState>(initialState = HomeState()) {
+
+    init {
+        updateState {
+            copy(
+                mapBoxPublicKey = buildConfigProvider.mapBoxPublicKey,
+                mapBoxStyleUri = buildConfigProvider.mapBoxStyleURI
+            )
+        }
+
+        viewModelScope.launch {
+            locationProvider.fetchUpdates().collect() {
+                updateState { copy(currentLocation = it) }
+            }
+        }
+    }
+}
